Added comments
added comments to source code

+++ routes/messageboard-api.js
@@ -1,55 +1,77 @@
 "use strict";
 
-var bcrypt = require("bcrypt");
-var ObjectId = require("mongodb").ObjectId;
-var uniqid = require("uniqid");
+var bcrypt = require("bcrypt"); //used to encrypt delete password field
+var ObjectId = require("mongodb").ObjectId; //used to find item in mongodb by _id
+var uniqid = require("uniqid"); //used to generate unique _id field for replies
 
-const SALT_ROUNDS = 10;
+const SALT_ROUNDS = 10; //used by bcrypt to generate hash
 
 module.exports = function(app, db) {
   const DB_TABLE = db.collection("board");
-  
+  /*DB "board" collection structure:
+      _id : thread id (auto generated)
+      board: String - board name
+      text: String - thread text
+      delete_password: - String (encrypted) - password to delete thread
+      created_on - String(Date)
+      bumped_on - String(Date) - DateTime when thread last modified
+      reported - boolean
+      replies - Array:
+                  _id - String - reply id generated using uniqid
+                  text - String - reply text
+                  delete_password - String(encrypted) - password to delete reply
+                  created_on - String(Date)
+                  reported - boolean
+  */
+
+  //starting page for messageboard app
   app.route("/messageboard").get(function(request, response) {
     response.sendFile(process.cwd() + "/dist/messageboard.html");
   });
 
+  //page for individual board
   app.route("/messageboard/:board/").get(function(req, res) {
     res.sendFile(process.cwd() + "/dist/messageboard-board.html");
   });
-  app.route("/messageboard/r/:postid").get(function(req, res) {
+
+  //page for individual thread
+  app.route("/messageboard/r/:threadid").get(function(req, res) {
     res.sendFile(process.cwd() + "/dist/messageboard-thread.html");
   });
 
+  //thread route/endpoints
   app.route("/messageboard/api/threads/:board")
+
+    //GET - get latest ten most recent threads for a given board and include the three most recent replies in order of last bumped/changed for each thread
     .get(function(req, res) {
       var board = req.params.board;
       //console.log("board:", board);
 
       DB_TABLE
         .find({ board: board })
-        .sort({ bumped_on: -1 })
-        .limit(10)
+        .sort({ bumped_on: -1 }) //sorted by bumped_on field from latest to oldest
+        .limit(10) //only show ten most recent threads
         .toArray((err, results) => {
           //console.log("results", results);
           let out = results.map(d => {
             //console.log("d", d);
-            d.replies.sort((a, b) => {
+            d.replies.sort((a, b) => { //sort replies in order of last bumped/changed
               a = new Date(a.bumped_on);
               b = new Date(b.bumped_on);
-              return a > b ? -1 : a < b ? 1 : 0;
+              return a > b ? -1 : a < b ? 1 : 0; //oldest first, newest last
             });
             let replycount = d.replies.length;
             let replies;
-            if (replycount > 3) {
+            if (replycount > 3) { //limit replies to the three most recent bumped repleis
               replies = [
-                d.replies[replycount - 1],
-                d.replies[replycount - 2],
-                d.replies[replycount - 3]
+                d.replies[replycount - 1], //last
+                d.replies[replycount - 2], //2nd from last
+                d.replies[replycount - 3]  //3rd from last
               ];
-            } else {
+            } else { //if 3 or less replies, include all replies from most recent to oldest
               replies = d.replies.reverse();
             }
-            replies = replies.map(r =>{
+            replies = replies.map(r =>{ //only return desired data fields for each reply
               return {
                 _id: r._id,
                 text: r.text,
@@ -57,7 +79,7 @@ module.exports = function(app, db) {
                 reported: r.reported
               }
             });
-            return {
+            return { //only returned desired fields for each thread
               _id: d._id,
               board: d.board,
               text: d.text,
@@ -68,21 +90,22 @@ module.exports = function(app, db) {
               reported: d.reported
             };
           });
-          res.json(out);
+          res.json(out); //output array of threads
         });
     })
 
+    //POST - create a thread from given board, text, and pass values; and return newThread fields
     .post(function(req, res) {
-      var board = req.params.board;
+      var board = req.params.board; //get board from url path
       //console.log("board", board);
-      var text = req.body.text;
+      var text = req.body.text; //get text and pass from form body
       var pass = req.body.delete_password;
       //console.log("req.body", req.body);
 
-      let newThread = {
+      let newThread = { //database fields for thread
         board: board,
         text: text,
-        delete_password: bcrypt.hashSync(pass, SALT_ROUNDS),
+        delete_password: bcrypt.hashSync(pass, SALT_ROUNDS), //encrypt delete_password before storing in database
         created_on: new Date(),
         bumped_on: new Date(),
         reported: false,
@@ -92,8 +115,8 @@ module.exports = function(app, db) {
       DB_TABLE.insertOne(newThread, (err, document) => {
         if (err) throw err;
         //console.log("document", document.ops[0]._id);
-        res.json({
-          _id: document.ops[0]._id,
+        res.json({ //return new thread values
+          _id: document.ops[0]._id, //get _id from database response
           board: newThread.board,
           text: newThread.text,
           created_on: newThread.created_on,
@@ -104,41 +127,42 @@ module.exports = function(app, db) {
       });
     })
 
+    //PUT - update the reported value on a given thread respond with message of if update was successful
     .put(function(req, res) {
-      //console.log("req.body:", req.body);
       var _id = req.body.thread_id;
       //console.log("thread_id", _id);
       let criteria;
       try {
-        criteria = { _id: ObjectId(_id) };
+        criteria = { _id: ObjectId(_id) }; //convert _id to database compatible _id field
       } catch (e) {
-        res.send("_id error");
+        res.send("_id error"); //send error message if incorrect _id format
         return;
       }
-      if (!criteria) return;
-      let update = { $set: { reported: true } };
+      if (!criteria) return; //stop if error found
+      let update = { $set: { reported: true } }; //set reported field on thread to true
 
       DB_TABLE.findOneAndUpdate(
         criteria,
         update,
         (err, result) => {
           //console.log("result:", result);
-          if (err || result.lastErrorObject.n == 0) {
+          if (err || result.lastErrorObject.n == 0) { //if error or no value was updated, send error message
             res.send("could not update " + _id);
             return;
           }
-          res.send("Thread reported successfully");
+          res.send("Thread reported successfully"); //success message
         }
       );
     })
 
+    //DELETE - delete a given thread from database
     .delete(function(req, res) {
-      var _id = req.body.thread_id;
+      var _id = req.body.thread_id; //_id and pass taken from form body
       var pass = req.body.delete_password;
 
       let criteria;
       try {
-        criteria = { _id: ObjectId(_id) };
+        criteria = { _id: ObjectId(_id) }; //convert _id to database compatible _id field
       } catch (e) {
         res.send("_id error");
         return;
@@ -149,12 +173,12 @@ module.exports = function(app, db) {
 
       DB_TABLE.findOne(criteria, (err, document) => {
         //console.log("document:", document);
-        if (err || document == [] || document == null) {
+        if (err || document == [] || document == null) { //if thread not found, rerturn failure message
           res.send("could not delete " + _id);
           return;
         }
 
-        if (bcrypt.compareSync(pass, document.delete_password)) {
+        if (bcrypt.compareSync(pass, document.delete_password)) { //unencrypt delete_passward and compare with pass given, if match -> delete thread
           //remove from database
           //console.log("TRUE!");
           DB_TABLE.deleteOne(
@@ -172,9 +196,12 @@ module.exports = function(app, db) {
       });
     }); //
 
+  //Reply route/endpoints
   app.route("/messageboard/api/replies/:thread_id")
+
+    //GET - get thread and it's replies for a given thread_id, return thread data
     .get(function(req, res) {
-      var thread_id = req.params.thread_id;
+      var thread_id = req.params.thread_id; //thread _id taken from url path
       //console.log("req.query:", req.query);
       //console.log("req.params:", req.params);
       //console.log("thread_id:", thread_id);
@@ -194,7 +221,7 @@ module.exports = function(app, db) {
           res.send("could not find " + thread_id);
           return;
         }
-        let replies = result.replies.map(d => {
+        let replies = result.replies.map(d => { //for each reply found only return specified values
           return {
             _id: d._id,
             text: d.text,
@@ -202,7 +229,7 @@ module.exports = function(app, db) {
             reported: d.reported
           };
         });
-        let out = {
+        let out = { //return thread values; replies are ordered from newest to oldest
           _id: result._id,
           board: result.board,
           created_on: result.created_on,
@@ -215,6 +242,7 @@ module.exports = function(app, db) {
       });
     })
 
+    //POST - create reply in a given thread; return reply data
     .post(function(req, res) {
       var thread_id = req.params.thread_id;
       //console.log("thread_id:", thread_id);
@@ -223,9 +251,9 @@ module.exports = function(app, db) {
       //console.log("req.body", req.body);
 
       let newReply = {
-        _id: uniqid(),
+        _id: uniqid(), //create a unique id value for each reply
         text: text,
-        delete_password: bcrypt.hashSync(pass, SALT_ROUNDS),
+        delete_password: bcrypt.hashSync(pass, SALT_ROUNDS), //encrypt delete_password
         created_on: new Date(),
         reported: false
       };
@@ -239,8 +267,8 @@ module.exports = function(app, db) {
       if (!criteria) return;
 
       let modify = {
-        $push: { replies: newReply },
-        $set: { bumped_on: new Date() }
+        $push: { replies: newReply }, //add new reply to end of replies array
+        $set: { bumped_on: new Date() } //update thread bumped_on value to current DateTime
       };
 
       DB_TABLE.findOneAndUpdate(
@@ -252,7 +280,7 @@ module.exports = function(app, db) {
             return;
           }
           //console.log("document", document);
-          res.json({
+          res.json({ //return specified values for newly created reply
             _id: newReply._id,
             text: newReply.text,
             created_on: newReply.created_on,
@@ -262,20 +290,21 @@ module.exports = function(app, db) {
       );
     })
 
+    //PUT - update reported field on a given reply; return message on whether successful or not
     .put(function(req, res) {
       var reply_id = req.body.reply_id;
       var thread_id = req.params.thread_id;
 
       let criteria;
       try {
-        criteria = { _id: new ObjectId(thread_id), "replies._id": reply_id };
+        criteria = { _id: new ObjectId(thread_id), "replies._id": reply_id }; //search by thread _id field and reply_id field in replies array
       } catch (e) {
         res.send("_id error");
         return;
       }
       if (!criteria) return;
 
-      let update = { $set: { "replies.$.reported": true } };
+      let update = { $set: { "replies.$.reported": true } }; //set reported field in matched reply in replies array
       //console.log("criteria:", criteria);
 
       DB_TABLE.findOneAndUpdate(
@@ -292,6 +321,7 @@ module.exports = function(app, db) {
       );
     })
 
+    //DELETE - update a reply text to '[deleted]', return if sucessful or not
     .delete(function(req, res) {
       var reply_id = req.body.reply_id;
       var thread_id = req.params.thread_id;
@@ -311,35 +341,36 @@ module.exports = function(app, db) {
       //console.log("criteria:", criteria);
 
       /*
-    let criteria;
-    try{
-      criteria = {_id: new ObjectId(thread_id)};
-    } catch (e){
-      res.send("_id error");
-      return
-    }
-    if(!criteria) return;
-    
-    
-    //let update = {$pull: {replies: {_id: reply_id}}};
+        //code to delete reply from replies array instead of just updating it
+        let criteria;
+        try{
+          criteria = {_id: new ObjectId(thread_id)};
+        } catch (e){
+          res.send("_id error");
+          return
+        }
+        if(!criteria) return;
+        
+        
+        //let update = {$pull: {replies: {_id: reply_id}}};
     */
-      let update = { $set: { "replies.$.text": "[deleted]" } };
+      let update = { $set: { "replies.$.text": "[deleted]" } }; //update text value of reply in replies array to "[deleted]"
 
-      DB_TABLE.findOne(criteria, (err, document) => {
+      DB_TABLE.findOne(criteria, (err, document) => { //first find thread and reply
         //console.log("document:", document.replies);
 
-        let thread = document.replies.filter(t => t._id == reply_id)[0];
+        let reply = document.replies.filter(t => t._id == reply_id)[0]; //get reply in return thread
         //console.log("thread:", thread);
 
-        if (thread == null || thread == undefined) {
+        if (reply == null || reply == undefined) { //if no reply with given reply_id found, return error message
           res.send("could not find reply " + reply_id);
           return;
         }
 
-        if (bcrypt.compareSync(pass, thread.delete_password)) {
+        if (bcrypt.compareSync(pass, reply.delete_password)) { //unencrypt reply delete_passward
           //remove from database
           //console.log("TRUE!");
-          DB_TABLE.findOneAndUpdate(
+          DB_TABLE.findOneAndUpdate( //if given pass matches delete_pass, update reply
             criteria,
             update,
             (er, result) => {
